+++ certbot-dns-azure/certbot_dns_azure/_internal/__init__.py
@@ -1,0 +1,1 @@
+"""Internal implementation of `~certbot_dns_azure.dns_azure` plugin."""